+++ ClickHouse.Driver/TypeSettings.cs
@@ -2,9 +2,7 @@
 
 namespace ClickHouse.Driver;
 
-internal record struct TypeSettings(bool useBigDecimal, string timezone)
+internal record struct TypeSettings(bool useBigDecimal)
 {
-    public static string DefaultTimezone = DateTimeZoneProviders.Tzdb.GetSystemDefault().Id;
-
-    public static TypeSettings Default => new TypeSettings(true, DefaultTimezone);
+    public static TypeSettings Default => new TypeSettings(true);
 }